+++ src/main/java/org/elasticsearch/action/admin/cluster/settings/ClusterUpdateSettingsClusterStateUpdateRequest.java
@@ -19,7 +19,7 @@
 
 package org.elasticsearch.action.admin.cluster.settings;
 
-import org.elasticsearch.cluster.ack.ClusterStateUpdateRequest;
+import org.elasticsearch.action.support.master.ClusterStateUpdateRequest;
 import org.elasticsearch.common.settings.Settings;
 
 import java.util.Locale;
