+++ src/main/java/org/elasticsearch/action/admin/cluster/settings/ClusterUpdateSettingsClusterStateUpdateResponse.java
@@ -19,7 +19,7 @@
 
 package org.elasticsearch.action.admin.cluster.settings;
 
-import org.elasticsearch.cluster.ack.ClusterStateUpdateResponse;
+import org.elasticsearch.action.support.master.ClusterStateUpdateResponse;
 import org.elasticsearch.common.settings.Settings;
 
 /**
