+++ src/main/java/org/elasticsearch/action/admin/indices/close/CloseIndexClusterStateUpdateRequest.java
@@ -18,7 +18,7 @@
 
 package org.elasticsearch.action.admin.indices.close;
 
-import org.elasticsearch.cluster.ack.ClusterStateUpdateRequest;
+import org.elasticsearch.action.support.master.ClusterStateUpdateRequest;
 
 import java.util.Arrays;
 import java.util.Locale;
