+++ src/main/java/org/elasticsearch/action/support/master/ClusterStateUpdateListener.java
@@ -1,4 +1,4 @@
-package org.elasticsearch.cluster.ack;
+package org.elasticsearch.action.support.master;
 
 /**
  * Listener used for cluster state updates processing
