+++ src/main/java/org/elasticsearch/cluster/metadata/MetaDataClusterService.java
@@ -26,7 +26,7 @@ import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsClus
 import org.elasticsearch.cluster.AckedClusterStateUpdateTask;
 import org.elasticsearch.cluster.ClusterService;
 import org.elasticsearch.cluster.ClusterState;
-import org.elasticsearch.cluster.ack.ClusterStateUpdateListener;
+import org.elasticsearch.action.support.master.ClusterStateUpdateListener;
 import org.elasticsearch.cluster.block.ClusterBlocks;
 import org.elasticsearch.cluster.node.DiscoveryNode;
 import org.elasticsearch.cluster.routing.allocation.AllocationService;
